workspace/7_framework/fastapi/day10/requirements.txt
fastapi
uvicorn[standard]
pytest
httpx
redis
slowapi
fastapi-cache2[redis]
